refactor: split mergeRenovateConfig
Split mergeRenovateConfig and it's helper functions into another module in order to break a future circular dependency.
